dmaengine: move struct in interface to use peripheral term
dmaengine history has a non inclusive terminology of dmaengine slave, I
feel it is time to replace that.

This moves structures in dmaengine interface with replacement of slave
to peripheral which is an appropriate term for dmaengine peripheral
devices

Since the change of name can break users, the new names have been added
with old structs kept as macro define for new names. Once the users have
been migrated, these macros will be dropped.

Signed-off-by: Vinod Koul <vkoul@kernel.org>